Removed all circular dependencies
