+++ lib/Avatar/AvatarWithPopover.js
@@ -3,7 +3,8 @@ import PropTypes from 'prop-types';
 import cx from 'classnames';
 import Overlay from 'react-bootstrap/lib/Overlay';
 import Popover from '../PopoverWrapper';
-import { Avatar, Icon } from '..';
+import Avatar from '.';
+import Icon from '../Icon';
 
 /**
  * @usage
