+++ lib/Button/ButtonWithIcon.js
@@ -1,7 +1,8 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import cx from 'classnames';
-import { Button, Icon } from '..';
+import Button from '.';
+import Icon from '../Icon';
 
 const ButtonWithIcon = props => {
   const classes = cx(`button__wrapper ${props.className}`, {
