+++ lib/CollectionsTable/CollectionsTableCell.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import CollectionsTableCellContent from './CollectionsTableCellContent';
-import { defaultProps, propTypes } from '.';
+import { defaultProps, propTypes } from './propTypes';
 
 const CollectionsTableCell = ({
   children,
