+++ lib/CollectionsTable/CollectionsTableCellContent.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import { defaultProps, propTypes } from '.';
+import { defaultProps, propTypes } from './propTypes';
 
 const CollectionsTableCellContent = ({ children, allowOverflow }) => {
   const classNames = cx('collections-table__cell-content', {
