+++ lib/CollectionsTable/CollectionsTableHead.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { propTypes, defaultProps } from '.';
+import { propTypes, defaultProps } from './propTypes';
 
 const CollectionsTableHead = ({ children, className, ...props }) => (
   <div className={`collections-table__head ${className}`} {...props}>
