+++ lib/CollectionsTable/CollectionsTableHeading.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { propTypes, defaultProps } from '.';
+import { propTypes, defaultProps } from './propTypes';
 import CollectionsTableCellContent from './CollectionsTableCellContent';
 
 const CollectionsTableHeading = ({ children, className, ...props }) => (
