+++ lib/CollectionsTable/propTypes.js
@@ -1,0 +1,10 @@
+import PropTypes from 'prop-types';
+
+export const propTypes = {
+  children: PropTypes.node.isRequired,
+  className: PropTypes.string
+};
+
+export const defaultProps = {
+  className: ''
+};
