+++ lib/ConfirmationDropdown/ConfirmationDropdownContent.js
@@ -1,7 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Button, Dropdown, Icon } from '../index';
-import { GATHER_UI_DROPDOWN } from '../Dropdown';
+import Button from '../Button';
+import Icon from '../Icon';
+import { GATHER_UI_DROPDOWN } from '../Dropdown/consts';
+import DropdownContent from '../Dropdown/DropdownContent';
+import DropdownFooter from '../Dropdown/DropdownFooter';
 
 const ConfirmationDropdownContent = (
   {
@@ -27,10 +30,10 @@ const ConfirmationDropdownContent = (
     : [...sharedButtonProps, 'link-default'];
 
   return (
-    <Dropdown.Content {...props}>
+    <DropdownContent {...props}>
       {children}
 
-      <Dropdown.Footer>
+      <DropdownFooter>
         <Icon
           name="loader"
           className="confirmation-dropdown__loader"
@@ -65,8 +68,8 @@ const ConfirmationDropdownContent = (
             Cancel
           </Button>
         )}
-      </Dropdown.Footer>
-    </Dropdown.Content>
+      </DropdownFooter>
+    </DropdownContent>
   );
 };
