+++ lib/ConfirmationDropdown/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import { Dropdown } from '..';
+import Dropdown from '../Dropdown';
 import ConfirmationDropdownContent from './ConfirmationDropdownContent';
 
 class ConfirmationDropdown extends Component {
