+++ lib/ConfirmationModal/index.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Button, ProgressButton, Modal } from '../index';
+import Button from '../Button';
+import ProgressButton from '../ProgressButton';
+import Modal from '../Modal';
 
 const ConfirmationModal = ({
   title,
