+++ lib/Dropdown/DropdownAction.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import { GATHER_UI_DROPDOWN } from './index';
+import { GATHER_UI_DROPDOWN } from './consts';
 
 const DropdownAction = (
   {
