+++ lib/Dropdown/DropdownContent.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import cx from 'classnames';
 import PropTypes from 'prop-types';
-import { GATHER_UI_DROPDOWN } from './index';
+import { GATHER_UI_DROPDOWN } from './consts';
 
 class DropdownContent extends Component {
   getStyle = () => {
