+++ lib/Dropdown/DropdownTrigger.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import { Button, Icon } from '../index';
-import { GATHER_UI_DROPDOWN } from './index';
+import Button from '../Button';
+import Icon from '../Icon';
+import { GATHER_UI_DROPDOWN } from './consts';
 
 class DropdownTrigger extends Component {
   handleClick = () => {
