+++ lib/Dropdown/index.js
@@ -9,8 +9,7 @@ import DropdownTrigger from './DropdownTrigger';
 import BoundaryClickWatcher from '../BoundaryClickWatcher';
 import DropdownHeader from './DropdownHeader';
 import DropdownFooter from './DropdownFooter';
-
-export const GATHER_UI_DROPDOWN = 'GATHER_UI_DROPDOWN';
+import { GATHER_UI_DROPDOWN } from './consts';
 
 class Dropdown extends Component {
   static Action = DropdownAction;
