+++ lib/Figure/FigurePreview.js
@@ -1,6 +1,6 @@
 import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
-import { Icon } from '../index';
+import Icon from '../Icon';
 
 const FigurePreview = ({
   caption,
