+++ lib/Guideline/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import { Button } from '..';
+import Button from '../Button';
 
 class Guideline extends Component {
   static propTypes = {
