+++ lib/MentionsForm/MentionsFormInput.js
@@ -1,7 +1,9 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { MentionsInput, Mention } from 'react-mentions';
-import { Avatar, AvatarInformation, UserSearchDropdown } from '..';
+import Avatar from '../Avatar';
+import AvatarInformation from '../Avatar/AvatarInformation';
+import UserSearchDropdown from '../UserSearchDropdown';
 import { filterUsers } from '../helpers';
 
 class MentionsFormInput extends Component {
