+++ lib/Modal/withModalTrigger.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Button } from '..';
+import Button from '../Button';
 
 export default function(buttonProps) {
   class withModalTrigger extends Component {
