+++ lib/ProgressButton/index.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Button, Icon } from '..';
+import Button from '../Button';
+import Icon from '../Icon';
 
 class ProgressButton extends Component {
   static defaultProps = {
