+++ lib/Search/SearchBoundaryListener.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { BoundaryClickWatcher } from '..';
+import BoundaryClickWatcher from '../BoundaryClickWatcher';
 import { SearchContext } from './SearchProvider';
 
 const SearchBoundaryListener = props => (
