+++ lib/Search/SearchInput.js
@@ -1,7 +1,9 @@
 import React, { Component } from 'react';
 import cx from 'classnames';
 import PropTypes from 'prop-types';
-import { Icon, Button, ShortcutTrigger } from '..';
+import Button from '../Button';
+import Icon from '../Icon';
+import ShortcutTrigger from '../ShortcutTrigger';
 
 class SearchInput extends Component {
   state = {
