+++ lib/Search/ToggleFilter.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { CheckToggle } from '..';
+import CheckToggle from '../CheckToggle';
 
 const ToggleFilter = props => (
   <div className="search-options__toggle">
