+++ lib/SearchDropdown/SearchResults.js
@@ -1,7 +1,9 @@
 import React, { Component, Fragment } from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import { EventCodeWatcher, Avatar, AvatarInformation } from '..';
+import EventCodeWatcher from '../EventCodeWatcher';
+import Avatar from '../Avatar';
+import AvatarInformation from '../Avatar/AvatarInformation';
 
 class SearchResults extends Component {
   static propTypes = {
