+++ lib/SearchInput/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { FontAwesomeIcon } from '../index';
+import FontAwesomeIcon from '../FontAwesomeIcon';
 
 const SearchInputField = ({
   placeholder,
