+++ lib/SelectionBar/SelectionBarCancel.js
@@ -1,6 +1,7 @@
 import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
-import { ShortcutTrigger, Button } from '..';
+import ShortcutTrigger from '../ShortcutTrigger';
+import Button from '../Button';
 
 const SelectionBarCancel = ({ clearSelection }) => (
   <Fragment>
