+++ lib/SelectionBar/SelectionBarInformation.js
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
 import pluralize from 'pluralize';
-import { Button } from '..';
+import Button from '../Button';
 
 const SelectionBarInformation = props => (
   <Fragment>
