+++ lib/Tabs/TabsItem.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import Tabs, { GATHER_UI_TABS } from './index';
+import { GATHER_UI_TABS } from './consts';
+import TabOptions from './TabsOptions';
 
 const TabItem = ({ children, id, options }, context) => {
   const { activeTabId, setActiveTab } = context[GATHER_UI_TABS];
@@ -20,7 +21,7 @@ const TabItem = ({ children, id, options }, context) => {
       tabIndex={-1}
     >
       {children}
-      {tabIsActive && options && <Tabs.Options options={options} />}
+      {tabIsActive && options && <TabOptions options={options} />}
     </span>
   );
 };
