+++ lib/Tabs/consts.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export const GATHER_UI_TABS = 'GATHER_UI_TABS';
