+++ lib/Tabs/index.js
@@ -5,8 +5,7 @@ import TabItem from './TabsItem';
 import TabButton from './TabsButton';
 import TabOptions from './TabsOptions';
 import TabForm from './TabsForm';
-
-export const GATHER_UI_TABS = 'GATHER_UI_TABS';
+import { GATHER_UI_TABS } from './consts';
 
 class Tabs extends Component {
   static Item = TabItem;
