+++ lib/TagUserForm/Users/TagUserFormUser.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Icon, Button, TooltipWrapper } from '../..';
+import Icon from '../../Icon';
+import Button from '../../Button';
+import TooltipWrapper from '../../TooltipWrapper';
 
 const TagUserFormUser = props => {
   if (props.isLocked) {
