+++ lib/TagUserForm/Users/TagUserFormUsers.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import TagUserFormUser from './TagUserFormUser';
-import { UserSearchDropdown } from '../..';
+import UserSearchDropdown from '../../UserSearchDropdown';
 
 const TagUserFormUsers = props => {
   if (props.lockedUsers.length === 0 && props.addedUsers.length === 0) {
