+++ lib/UserList/CurrentUserList.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Button, Avatar, AvatarInformation } from '..';
+import Button from '../Button';
+import Avatar from '../Avatar';
+import AvatarInformation from '../Avatar/AvatarInformation';
 
 const CurrentUserList = props => {
   const {
