+++ lib/UserList/OtherUserList.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Avatar, AvatarInformation } from '..';
+import Avatar from '../Avatar';
+import AvatarInformation from '../Avatar/AvatarInformation';
 
 const OtherUserList = props => {
   const { otherUsers, otherUsersHeading } = props;
