+++ lib/UserList/index.js
@@ -1,7 +1,9 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import { Icon, Button, SearchDropdown } from '..';
+import Icon from '../Icon';
+import Button from '../Button';
+import SearchDropdown from '../SearchDropdown';
 import CurrentUserList from './CurrentUserList';
 import OtherUserList from './OtherUserList';
