+++ lib/UserSearch/UserSearchHead.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { CheckToggle } from '..';
+import CheckToggle from '../CheckToggle';
 
 const UserSearchHead = ({
   searchHeading,
