+++ lib/UserSearch/UserSearchList.js
@@ -1,6 +1,8 @@
 import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
-import { Dropdown, Avatar, AvatarInformation } from '..';
+import Dropdown from '../Dropdown';
+import Avatar from '../Avatar';
+import AvatarInformation from '../Avatar/AvatarInformation';
 import UserSearchListUserName from './UserSearchListUserName';
 
 const UserSearchList = ({
