+++ lib/UserSearch/UserSearchListUserName.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import { Icon } from '..';
+import Icon from '../Icon';
 
 const UserSearchListUserName = ({ name, isSelected }) => {
   const classes = cx('user-search__name', { 'is-active': isSelected });
