+++ lib/UserSearch/index.js
@@ -1,6 +1,6 @@
 import React, { Component, Fragment } from 'react';
 import PropTypes from 'prop-types';
-import { Dropdown } from '..';
+import Dropdown from '../Dropdown';
 import { filterUsers } from '../helpers';
 import UserSearchList from './UserSearchList';
 import UserSearchHead from './UserSearchHead';
