+++ lib/UserSearchDropdown/index.js
@@ -1,6 +1,9 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Dropdown, Icon, TooltipWrapper, UserSearch } from '..';
+import Dropdown from '../Dropdown';
+import Icon from '../Icon';
+import TooltipWrapper from '../TooltipWrapper';
+import UserSearch from '../UserSearch';
 
 class UserSearchDropdown extends Component {
   searchInputRef = React.createRef();
