+++ lib/images/ImageLoader/index.js
@@ -1,7 +1,7 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import { Icon } from '../..';
+import Icon from '../../Icon';
 
 class ImageLoader extends PureComponent {
   static propTypes = {
